+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - Added Changelog
 - Added fatrop support (requires casadi 3.6.6)
+- add "euler" to integrator for discretization options in multiple shooting